+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 
 ### Changed
 - A percentage of your peasant population is now invulnerable to fireball damage (scales down during the round from 90% on Day 4 to 70% on Day 44)
+- Up to 50% of your vulnerable peasant population is protected by defensive WPA
 
 ## [1.37.2] - 2023-10-24
 ### Added